# fix(web): disable fat-finger predictions for perf