@jahorton 2021-06-21
>     1. Implement sliding window for `buildTransformFrom` (not currently used, see point 3). But did improve performance by about 30msec, so thought I'd leave it in.

Actually, it's still used, even then.

https://github.com/keymanapp/keyman/blob/349e144e59d27746d1802282344cba67e0e26d2e/common/core/web/input-processor/src/text/inputProcessor.ts#L157-L159

https://github.com/keymanapp/keyman/blob/349e144e59d27746d1802282344cba67e0e26d2e/web/source/dom/domOverrides.ts#L19-L41

It's literally at the core of how embedded operates.