@jahorton 2021-06-25
So... now that #5319's been merged in, I think points 1 and 3 are now moot.  Point 2 (embedded `display: none`) sounds pretty useful though, so that's likely worth preserving, even if it'll need a bit more vetting first.