@mcdurdin 2021-06-25
> Point 2 (embedded `display: none`) sounds pretty useful though, so that's likely worth preserving, even if it'll need a bit more vetting first.

I've been continuing to test this with no side effects observed on Android. Have not tested on iOS. I'll be trying to setup a pair of PRs to tidy this up today, so we can close this off and get the release out.