@mcdurdin 2021-06-25
> I'll be trying to setup a pair of PRs to tidy this up today, so we can close this off and get the release out.

I've been diverted onto other fires so will not get to this until next week now.